Реорганизация buttons.*
