+++ server/entities/env_local.h
@@ -1,0 +1,36 @@
+/***
+*
+*	Copyright (c) 1996-2002, Valve LLC. All rights reserved.
+*	
+*	This product contains software technology licensed from Id 
+*	Software, Inc. ("Id Technology").  Id Technology (c) 1996 Id Software, Inc. 
+*	All Rights Reserved.
+*
+*   Use, distribution, and modification of this source code and/or resulting
+*   object code is restricted to non-commercial enhancements to products from
+*   Valve LLC.  All other use, distribution, or modification is prohibited
+*   without written permission from Valve LLC.
+*
+****/
+
+#pragma once
+
+#include "extdll.h"
+#include "util.h"
+#include "cbase.h"
+#include "client.h"
+#include "player.h"
+#include "doors.h"
+
+#define SF_LOCAL_START_ON		BIT( 0 )
+#define SF_LOCAL_REMOVE_ON_FIRE	BIT( 1 )
+
+class CEnvLocal : public CBaseDelay
+{
+	DECLARE_CLASS( CEnvLocal, CBaseDelay );
+public:
+	void Spawn( void );
+	void Think( void );
+	void Use( CBaseEntity *pActivator, CBaseEntity *pCaller, USE_TYPE useType, float value );
+	virtual int ObjectCaps( void ) { return BaseClass :: ObjectCaps() & ~FCAP_ACROSS_TRANSITION; }
+};
