+++ server/entities/env_time.h
@@ -1,0 +1,37 @@
+/***
+*
+*	Copyright (c) 1996-2002, Valve LLC. All rights reserved.
+*	
+*	This product contains software technology licensed from Id 
+*	Software, Inc. ("Id Technology").  Id Technology (c) 1996 Id Software, Inc. 
+*	All Rights Reserved.
+*
+*   Use, distribution, and modification of this source code and/or resulting
+*   object code is restricted to non-commercial enhancements to products from
+*   Valve LLC.  All other use, distribution, or modification is prohibited
+*   without written permission from Valve LLC.
+*
+****/
+
+#pragma once
+
+#include "extdll.h"
+#include "util.h"
+#include "cbase.h"
+#include "client.h"
+#include "player.h"
+#include "doors.h"
+
+class CEnvTime : public CBaseDelay
+{
+	DECLARE_CLASS( CEnvTime, CBaseDelay );
+public:
+	void Spawn( void );
+	void Think( void );
+	void Activate( void );
+	void KeyValue( KeyValueData *pkvd );
+	int ObjectCaps( void ) { return BaseClass :: ObjectCaps() & ~FCAP_ACROSS_TRANSITION; }
+private:
+	float m_flLevelTime;
+	Vector m_vecTime;
+};
